Version 2
This is version 2 of the original python file I uploaded. I made following changes to the original file:

(i) I am now using a loop to assign variables to each of the 8 'table' tags I scraped into "World_Cup" list.

(ii) After feature-engineering the contents of  the scraped <table>, <tbody>, <tr> and <td> tags, I am now immediately displaying the resulting list of dictionaries and the dataframe obtained. This is based on the feedback during my presentation of the work.

(iii) To avoid failing insert operations into my Mongodb, I added some lines of codes to remove or drop all collections if it already exists. I observe that insert operations fail because the documents already exist in the DB.